airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

yfinance
pandas
numpy
apache-airflow-providers-postgres==6.5.3
psycopg2-binary
SQLAlchemy